refactor: change <svg> calls to use new TS SVGs
